+++ packages/ethereum-indexer-js-processor/src/processor/JSObjectEventProcessor.ts
@@ -45,8 +45,6 @@ export class JSObjectEventProcessor<ABI extends Abi, ProcessResultType extends J
 	constructor(private singleEventProcessor: SingleEventJSONProcessor<ABI, ProcessResultType, ProcessorConfig>) {
 		this.version = singleEventProcessor.version;
 		const state = singleEventProcessor.createInitialState();
-		const initial_frozen = Object.isFrozen(state);
-		console.log({initial_frozen});
 		const history = {
 			blockHashes: {},
 			reversals: {},
@@ -57,9 +55,6 @@ export class JSObjectEventProcessor<ABI extends Abi, ProcessResultType extends J
 			history,
 		};
 		this.history = new History(history);
-
-		const constructor_frozen = Object.isFrozen(this._json.state);
-		console.log({constructor_frozen});
 	}
 
 	copyFrom(otherProcessor: JSObjectEventProcessor<ABI, ProcessResultType, ProcessorConfig>) {
@@ -94,8 +89,6 @@ export class JSObjectEventProcessor<ABI extends Abi, ProcessResultType extends J
 	async reset() {
 		namedLogger.info('JSObjectEventProcessor reseting...');
 		const state = this.singleEventProcessor.createInitialState();
-		const reset_frozen = Object.isFrozen(state);
-		console.log({reset_frozen});
 		this.version = this.singleEventProcessor.version;
 		const history = {
 			blockHashes: {},
@@ -149,15 +142,9 @@ export class JSObjectEventProcessor<ABI extends Abi, ProcessResultType extends J
 
 					this._json.state = state;
 					this._json.lastSync = lastSyncFromExistingState;
-
-					const fetched_frozen = Object.isFrozen(this._json.state);
-					console.log({fetched_frozen});
 				}
 			}
 		}
-
-		const loaded_frozen = Object.isFrozen(this._json.state);
-		console.log({loaded_frozen});
 
 		if (!this._json.lastSync) {
 			return undefined;
@@ -190,9 +177,6 @@ export class JSObjectEventProcessor<ABI extends Abi, ProcessResultType extends J
 					if (!lastBlockDeleted || event.blockHash != lastBlockDeleted) {
 						namedLogger.info(`JSObjectEventProcessor preparing block...`);
 						this._json.state = this.history.reverseBlock(event.blockNumber, event.blockHash, this._json.state);
-
-						const removed_frozen = Object.isFrozen(this._json.state);
-						console.log({removed_frozen});
 						lastBlockDeleted = event.blockHash;
 					}
 				} else {
@@ -205,8 +189,6 @@ export class JSObjectEventProcessor<ABI extends Abi, ProcessResultType extends J
 									this.history.setReversal(reversePatches);
 								}) as ProcessResultType;
 								this._json.state = finalizedDraft as unknown as ProcessResultType;
-								const finalized_frozen = Object.isFrozen(this._json.state);
-								console.log({finalized_frozen});
 							}
 
 							this.history.setBlock(event.blockNumber, event.blockHash);
@@ -226,9 +208,6 @@ export class JSObjectEventProcessor<ABI extends Abi, ProcessResultType extends J
 					this.history.setReversal(reversePatches);
 				}) as ProcessResultType;
 				this._json.state = finalizedDraft;
-
-				const final_finalized_frozen = Object.isFrozen(this._json.state);
-				console.log({final_finalized_frozen});
 			}
 
 			let lastLastSync;