+++ examples/event-processor-conquest-fplay/CHANGELOG.md
@@ -1,5 +1,12 @@
 # event-processor-conquest-fplay
 
+## 0.0.51
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer-js-processor@0.6.25
+
 ## 0.0.50
 
 ### Patch Changes