+++ examples/web-demo/CHANGELOG.md
@@ -1,5 +1,15 @@
 # web-demo
 
+## 0.1.57
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer-browser@0.6.27
+  - event-processor-bleeps@0.0.51
+  - event-processor-conquest-eth@0.0.51
+  - event-processor-nfts@0.0.51
+
 ## 0.1.56
 
 ### Patch Changes